show diagnostic output with annotations
We are writing a new tool to show diagnostic output with annotations.
The tool will be called `cbordiag` and will be used to parse and
display diagnostic information from CBOR (Concise Binary Object
Representation) data.  We'll not be using the go-cbor library for this
project, but instead will be writing our own parser to generate the
annotated output.  We can reference existing annotated diagnostic parsers
in Ruby and Rust for inspiration.

Next steps:
- write test cases for each function in cbordiag.go
- always return the complete file -- do not stub, summarize, or replace
  code with comments